+++ src/extension.js
@@ -5,22 +5,14 @@ const { v4: uuidv4 } = require('uuid');
 const cheerio = require("cheerio");
 var Walk = require("@root/walk");
 
-const hdr = require("./htr.js");
+const htr = require("./htr.js");
+const utils = require("./utils.js");
+const langs = require("./langs.js");
 
 // load root webview document
 var $ = cheerio.load(fs.readFileSync(path.join(__dirname, 'webview.html'), { encoding: 'utf8' }))
 
 let settings = vscode.workspace.getConfiguration('draw');
-
-function getNonce() {
-  let text = '';
-  const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  for (let i = 0; i < 32; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
-}
-
 
 // inject js/css into base html document
 function loadWebviewFiles(err, pathname, dirent) {
@@ -38,13 +30,10 @@ function loadWebviewFiles(err, pathname, dirent) {
   return Promise.resolve();
 }
 
-var webviewContent;
-Walk.walk(path.join(__dirname, "webview"), loadWebviewFiles).then(function () {
-  webviewContent = $.root().html().replace(/ToBeReplacedByRandomToken/g, getNonce())
-});
+Walk.walk(path.join(__dirname, "webview"), loadWebviewFiles).then(function () { });
 
 /** @param {vscode.ExtensionContext} context */
-function activate(context) {
+exports.activate = function (context) {
 
   // values for webview status
   /** @type {vscode.WebviewPanel | undefined} */
@@ -68,8 +57,18 @@ function activate(context) {
         enableScripts: true
       }
     );
+    const nonce = utils.nonce()
 
-    currentPanel.webview.html = getWebviewContent();
+    // append script at path to the document body
+    function inject(path) {
+      const toolkitUri = currentPanel.webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, ...[path]))
+      $("body").append(`<script type="module" nonce="${nonce}" src="${toolkitUri}"></script>`)
+    }
+
+    inject("./node_modules/@vscode/webview-ui-toolkit/dist/toolkit.js")
+    inject("./node_modules/iink-js/dist/iink.min.js")
+
+    currentPanel.webview.html = $.root().html().replace(/ToBeReplacedByRandomToken/g, nonce)
     // Handle messages from the webview
     currentPanel.webview.onDidReceiveMessage(
       message => {
@@ -116,10 +115,6 @@ function activate(context) {
     );
   }
 
-  function showPanel() {
-    currentPanel.reveal();
-  }
-
   function getEditorText(show) {
     let currentEditor_ = currentEditor
     let currentLine_ = currentLine
@@ -164,7 +159,7 @@ function activate(context) {
         currentEditor = currentEditor_
         currentLine = currentLine_
         if (settings.directory) {
-          let link = readLink(currentEditor.document.languageId, text)
+          let link = langs.readLink(currentEditor.document.languageId, text)
           if (link) text = fs.readFileSync(path.join(vscode.workspace.rootPath, link), { encoding: 'utf-8' })
         }
         if (topush) {
@@ -174,73 +169,6 @@ function activate(context) {
     }, 100)
   }
 
-  // write text to filename
-  function writeFile(text, filename) {
-    dir = path.join(vscode.workspace.rootPath, settings.directory);
-    if (!fs.existsSync(dir)) {
-      fs.mkdirSync(dir);
-    }
-    fs.writeFileSync(path.join(dir, filename), text, { encoding: 'utf8' });
-  }
-
-  // return language formatted link to filename; optionally, with alt text
-  function createLink(language, filename, alt) {
-    // https://hyperpolyglot.org/lightweight-markup
-    switch (language) {
-
-      case 'markdown':
-        return `![${alt}](${filename})`
-
-      case 'asciidoc':
-        return `image::${filename}[${alt}]`
-
-      case 'restructuredtext':
-        return `.. image:: ${filename}` // TODO: add alt text `\n   :alt: ${alt}`
-
-      // case 'mediawiki':
-      //   return `[[File:${filename}|alt=${alt}]]`
-
-      // case 'org':
-      //   return `[[${filename}]]`
-
-    }
-  }
-
-  Array.prototype.swap = function (a, b) {
-    this[a] = this.splice(b, 1, this[a])[0]
-    return this
-  }
-
-  // return alt text and filename (in that order) from link in language format
-  function readLink(language, link) {
-    // https://hyperpolyglot.org/lightweight-markup
-    let match;
-    switch (language) {
-
-      case 'markdown':
-        match = link.match(/!\[(.*)\]\((.*)\)/)
-        break
-
-      case 'asciidoc':
-        match = link.match(/image::(.*)\[(.*)\]/)
-        if (match) match.swap(1, 2)
-        break
-
-      case 'restructuredtext':
-        // TODO: support multiline text for alt
-        match = link.match(/..() image:: (.*)/)
-        break
-
-      // case 'mediawiki':
-      //   match = link.match()
-      //   break
-
-      // case 'org':
-      //   match = link.match()
-      //   break
-    }
-    if (match) return match[1], match[2]
-  }
 
   function setEditorText(text, control) {
     if (settings.directory && !text.startsWith("$$")) {
@@ -253,36 +181,33 @@ function activate(context) {
         filename = match[2].replace(/^.*[\\\/]/, '')
       }
 
-      writeFile(text, filename)
+      utils.write(text, filename)
       let name = path.join(settings.directory, filename)
-      text = createLink(currentEditor.document.languageId, name, alt)
+      text = langs.createLink(currentEditor.document.languageId, name, alt)
     }
 
     if (!currentEditor || currentEditor.document.isClosed) {
       vscode.window.showErrorMessage('The text editor has been closed');
       return;
     }
+
     let p = vscode.window.showTextDocument(currentEditor.document, {
       viewColumn: currentEditor.viewColumn,
       selection: new vscode.Range(currentLine, 0, currentLine, 0)
-    })
-      .then((editor) => editor.edit(edit => {
-        let lf = '\n'
-        edit.replace(new vscode.Range(currentLine, 0, currentLine + 1, 0), text + lf);
-        resetCheckStrings(text.split('\n')[0] + '\n')
-      }))
+    }).then((editor) => editor.edit(edit => {
+      edit.replace(new vscode.Range(currentLine, 0, currentLine + 1, 0), text + '\n');
+      resetCheckStrings(text.split('\n')[0] + '\n')
+    }))
+
     if (control !== 0) {
-      p = p
-        .then(() => vscode.window.showTextDocument(currentEditor.document, {
+      p = p.then(() => {
+        vscode.window.showTextDocument(currentEditor.document, {
           viewColumn: currentEditor.viewColumn,
           selection: new vscode.Range(currentLine + control, 0, currentLine + control, 0)
-        })) // the next line somehow not working, so use this line
-        // .then(() => currentEditor.revealRange(
-        //   new vscode.Range(currentLine + control, 0, currentLine + control, 0)
-        // )) 
-        .then(() => {
-          pushCurrentLine()
         })
+      }).then(() => {
+        pushCurrentLine()
+      })
     }
   }
 
@@ -307,24 +232,18 @@ function activate(context) {
   context.subscriptions.push(
     vscode.commands.registerCommand('draw.editCurrentLineAsSVG', () => {
       if (currentPanel) {
-        showPanel()
+        currentPanel.reveal();
         pushCurrentLine()
       } else {
-        vscode.commands.executeCommand('workbench.action.editorLayoutTwoRowsRight')
-          .then(() => {
-            createNewPanel()
-            pushCurrentLine()
-          })
+        vscode.commands.executeCommand('workbench.action.editorLayoutTwoRowsRight').then(() => {
+          createNewPanel()
+          pushCurrentLine()
+        })
       }
     }),
-    vscode.commands.registerCommand('draw.configureHDR', () => {
-      hdr.init(context, currentPanel);
+    vscode.commands.registerCommand('draw.configureHTR', () => {
+      htr.init(context, currentPanel);
     })
   );
 
-}
-exports.activate = activate;
-
-function getWebviewContent() {
-  return webviewContent
 }